Adds BaseBinner and refactors Cond. Inf. Tree
